@@ -5,17 +5,18 @@
 
 	<artifactId>de.claudioaltamura</artifactId>
     <groupId>keycloak-spring-boot-rest</groupId>
-	<version>0.0.1-SNAPSHOT</version>
-	<name>Keycloak Spring Boot REST</name>
+	<version>0.0.2-SNAPSHOT</version>
+	<name>Keycloak Spring Boot REST example</name>
 	<description>Keycloak Spring Boot REST example</description>
 
     <properties>
       <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
       <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-      <maven.compiler.source>11</maven.compiler.source>
-      <maven.compiler.target>11</maven.compiler.target>
-      <spring-boot.version>2.6.0</spring-boot.version>
-      <keycloak-bom.version>18.0.0</keycloak-bom.version>
+      <maven.compiler.source>17</maven.compiler.source>
+      <maven.compiler.target>17</maven.compiler.target>
+      <spring-boot.version>3.0.2</spring-boot.version>
+      <keycloak-bom.version>20.0.3</keycloak-bom.version>
+        <spotless.version>2.33.0</spotless.version>
     </properties>
 
     <dependencyManagement>
@@ -78,12 +79,6 @@
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-test</artifactId>
             <scope>test</scope>
-            <exclusions>
-                <exclusion>
-                    <groupId>org.junit.vintage</groupId>
-                    <artifactId>junit-vintage-engine</artifactId>
-                </exclusion>
-            </exclusions>
         </dependency>
 
         <dependency>
@@ -96,6 +91,20 @@
 
     <build>
         <plugins>
+            <plugin>
+                <groupId>com.diffplug.spotless</groupId>
+                <artifactId>spotless-maven-plugin</artifactId>
+                <version>${spotless.version}</version>
+                <executions>
+                    <execution>
+                        <goals>
+                            <goal>apply</goal>
+                        </goals>
+                        <phase>process-sources</phase>
+                    </execution>
+                </executions>
+            </plugin>
+
             <plugin>
                 <groupId>org.springframework.boot</groupId>
                 <artifactId>spring-boot-maven-plugin</artifactId>
